requirements.txt
anthropic>=0.40.0
openai>=1.50.0
qdrant-client>=1.12.0
python-frontmatter>=1.1.0
PyYAML>=6.0.2
watchdog>=5.0.0
pytest>=8.0.0
